+++ number/mod.test.ts
@@ -40,6 +40,26 @@ Deno.test({
 });
 
 Deno.test({
+  name: "metersToFeet works",
+  fn: () => {
+    assertEquals(
+      NumberUtils.roundTo(NumberUtils.metersToFeet(5), 6),
+      16.4042,
+    );
+  },
+});
+
+Deno.test({
+  name: "feetToMeters works",
+  fn: () => {
+    assertEquals(
+      NumberUtils.roundTo(NumberUtils.feetToMeters(5), 6),
+      1.524,
+    );
+  },
+});
+
+Deno.test({
   name: "isEven works",
   fn: () => {
     assertEquals(NumberUtils.isEven(2), true);