+++ geo/CHANGELOG.md
@@ -1,5 +1,10 @@
 # `geo-utils` CHANGELOG
 
+## 1.1.0
+
+- New functions:
+  - `feetToMeters()` and `metersToFeet()`
+
 ## 1.0.0
 
 - Initial release.