+++ number/CHANGELOG.md
@@ -1,5 +1,10 @@
 # `number-utils` CHANGELOG
 
+## 1.2.0
+
+- New functions:
+  - `feetToMeters()` and `metersToFeet()`
+
 ## 1.1.0
 
 - New functions: